pyproject.toml
[tool.poetry]
name = "data-helper"
version = "0.1.0"
description = ""
authors = ["Joseph Machado"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
aiohttp = "3.9.5"
aiosignal = "1.3.1"
annotated-types = "0.7.0"
anyio = "4.3.0"
async-timeout = "4.0.3"
attrs = "23.2.0"
beautifulsoup4 = "4.12.3"
blis = "0.7.11"
catalogue = "2.0.10"
certifi = "2024.2.2"
charset-normalizer = "3.3.2"
click = "8.1.7"
cloudpathlib = "0.16.0"
confection = "0.1.4"
cymem = "2.0.8"
dataclasses-json = "0.6.6"
deprecated = "1.2.14"
dirtyjson = "1.0.8"
distro = "1.9.0"
exceptiongroup = "1.2.1"
frozenlist = "1.4.1"
fsspec = "2024.5.0"
greenlet = "3.0.3"
h11 = "0.14.0"
httpcore = "1.0.5"
httpx = "0.27.0"
idna = "3.7"
jinja2 = "3.1.4"
joblib = "1.4.2"
jsonpath-ng = "1.6.1"
langcodes = "3.4.0"
language-data = "1.2.0"
llama-index = "0.10.37"
llama-index-agent-openai = "0.2.5"
llama-index-cli = "0.1.12"
llama-index-core = "0.10.37.post1"
llama-index-embeddings-openai = "0.1.10"
llama-index-indices-managed-llama-cloud = "0.1.6"
llama-index-legacy = "0.9.48"
llama-index-llms-openai = "0.1.20"
llama-index-multi-modal-llms-openai = "0.1.6"
llama-index-program-openai = "0.1.6"
llama-index-question-gen-openai = "0.1.3"
llama-index-readers-file = "0.1.22"
llama-index-readers-llama-parse = "0.1.4"
llama-parse = "0.4.3"
llamaindex-py-client = "0.1.19"
marisa-trie = "1.1.1"
markupsafe = "2.1.5"
marshmallow = "3.21.2"
multidict = "6.0.5"
murmurhash = "1.0.10"
mypy-extensions = "1.0.0"
nest-asyncio = "1.6.0"
networkx = "3.3"
nltk = "3.8.1"
numpy = "1.26.4"
openai = "1.30.1"
packaging = "24.0"
pandas = "2.2.2"
pillow = "10.3.0"
ply = "3.11"
preshed = "3.0.9"
pydantic = "2.7.1"
pydantic-core = "2.18.2"
pypdf = "4.2.0"
python-dateutil = "2.9.0.post0"
pytz = "2024.1"
pyyaml = "6.0.1"
regex = "2024.5.15"
requests = "2.32.2"
six = "1.16.0"
smart-open = "6.4.0"
sniffio = "1.3.1"
soupsieve = "2.5"
spacy = "3.7.4"
spacy-legacy = "3.0.12"
spacy-loggers = "1.0.5"
sqlalchemy = ">=1.4.36,<2.0"
srsly = "2.4.8"
striprtf = "0.0.26"
tenacity = "8.3.0"
thinc = "8.2.3"
tiktoken = "0.7.0"
tqdm = "4.66.4"
typer = "0.9.4"
typing-inspect = "0.9.0"
typing-extensions = "4.11.0"
tzdata = "2024.1"
urllib3 = "2.2.1"
wasabi = "1.1.2"
weasel = "0.3.4"
wrapt = "1.16.0"
yarl = "1.9.4"
black = "^24.4.2"
mypy = "^1.10.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
acryl-datahub = "^0.13.3.3"
datasets = "^2.20.0"
pyarrow = "^15.0.0"
fastavro = "^1.4.7"
apache-airflow = "^2.9.3"
grafana_api = "^1.0.0"
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
